refactor: :recycle: move get_token to own file
