+++ src/zen_do/cli.py
@@ -1,9 +1,6 @@
-import os
-
-import keyring
 from cyclopts import App
-from keyring.errors import NoKeyringError
 
+from zen_do.get_token import get_token
 from zen_do.zenodo import (
     zenodo_create_record,
     zenodo_get_record,
@@ -29,34 +26,3 @@ def zenodo_publish(sandbox: bool = False) -> None:
     else:
         zenodo_create_record(token)
         print("New Zenodo record created successfully!")
-
-
-SERVICE_NAME = "zen-do"
-
-
-def get_token(sandbox: bool = False) -> str:
-    """Gets the Zenodo token from the system keyring or environment variables.
-
-    Args:
-        sandbox: Whether to get the token for the Zenodo sandbox environment.
-
-    Returns:
-        The token.
-
-    Raises:
-        RuntimeError: If no token is found.
-    """
-    token_name = "ZENODO_SANDBOX_TOKEN" if sandbox else "ZENODO_TOKEN"
-    try:
-        token = keyring.get_password(service_name=SERVICE_NAME, username=token_name)
-    except NoKeyringError:
-        token = None
-
-    token = token or os.getenv(token_name)
-
-    if not token:
-        raise RuntimeError(
-            f"No value found for {token_name!r} in the system "
-            "keyring or the environment variables."
-        )
-    return token
